+++ src/configs/extra/cli.ts
@@ -1,4 +1,4 @@
-import {GLOB_JS_TS_X_EXTENSION} from '../../constants';
+import {ERROR, GLOB_JS_TS_X_EXTENSION} from '../../constants';
 import {type UnConfigOptions, createConfigBuilder} from '../../eslint';
 import {assignDefaults} from '../../utils';
 import type {UnConfigFn} from '../index';
@@ -35,6 +35,7 @@ export const cliUnConfig: UnConfigFn<'cli'> = (context) => {
     .disableAnyRule('unicorn', 'no-process-exit')
     .disableAnyRule('', 'no-console')
     .disableAnyRule('import', 'no-extraneous-dependencies')
+    .addAnyRule('unicorn', 'prefer-top-level-await', ERROR)
     .addOverrides();
 
   return {